requirements.txt
discord.py>=2.3.0
python-dateutil>=2.8.2
aiosqlite>=0.19.0
